requirements.txt
discord.py
python-dotenv
ipykernel
discord.py[voice]
yt-dlp